+++ src/GitHub_Updater/GitHub_API.php
@@ -25,7 +25,7 @@ if ( ! defined( 'WPINC' ) ) {
  * @package Fragen\GitHub_Updater
  * @author  Andy Fragen
  */
-class GitHub_API extends API {
+class GitHub_API extends API implements API_Interface {
 
 	/**
 	 * Holds loose class method name.
@@ -47,7 +47,7 @@ class GitHub_API extends API {
 	/**
 	 * Read the remote file and parse headers.
 	 *
-	 * @param $file
+	 * @param string $file Filename.
 	 *
 	 * @return bool
 	 */
@@ -114,7 +114,7 @@ class GitHub_API extends API {
 	/**
 	 * Read the remote CHANGES.md file.
 	 *
-	 * @param $changes
+	 * @param string $changes Changelog filename.
 	 *
 	 * @return bool
 	 */
@@ -269,7 +269,7 @@ class GitHub_API extends API {
 	}
 
 	/**
-	 * Construct $this->type->download_link using Repository Contents API
+	 * Construct $this->type->download_link using Repository Contents API.
 	 * @url http://developer.github.com/v3/repos/contents/#get-archive-link
 	 *
 	 * @param boolean $rollback      for theme rollback
@@ -322,12 +322,12 @@ class GitHub_API extends API {
 	/**
 	 * Add appropriate access token to endpoint.
 	 *
-	 * @param $git
-	 * @param $endpoint
+	 * @param object $git
+	 * @param string $endpoint
 	 *
 	 * @access private
 	 *
-	 * @return string
+	 * @return string $endpoint
 	 */
 	private function add_access_token_endpoint( $git, $endpoint ) {
 		// This will return if checking during shiny updates.
@@ -360,12 +360,12 @@ class GitHub_API extends API {
 	/**
 	 * Create GitHub API endpoints.
 	 *
-	 * @param $git      object
-	 * @param $endpoint string
+	 * @param object $git
+	 * @param string $endpoint
 	 *
 	 * @return string $endpoint
 	 */
-	protected function add_endpoints( $git, $endpoint ) {
+	public function add_endpoints( $git, $endpoint ) {
 		switch ( $git::$method ) {
 			case 'file':
 			case 'readme':
@@ -417,7 +417,7 @@ class GitHub_API extends API {
 	 *
 	 * @return object|array $arr Array of tag numbers, object is error.
 	 */
-	protected function parse_tag_response( $response ) {
+	public function parse_tag_response( $response ) {
 		if ( isset( $response->message ) ) {
 			return $response;
 		}
@@ -439,7 +439,7 @@ class GitHub_API extends API {
 	 *
 	 * @return array $arr Array of meta variables.
 	 */
-	protected function parse_meta_response( $response ) {
+	public function parse_meta_response( $response ) {
 		$arr      = array();
 		$response = array( $response );
 
@@ -461,7 +461,7 @@ class GitHub_API extends API {
 	 *
 	 * @return array $arr Array of changes in base64.
 	 */
-	protected function parse_changelog_response( $response ) {
+	public function parse_changelog_response( $response ) {
 		$arr      = array();
 		$response = array( $response );