+++ src/GitHub_Updater/GitLab_API.php
@@ -25,7 +25,7 @@ if ( ! defined( 'WPINC' ) ) {
  * @package Fragen\GitHub_Updater
  * @author  Andy Fragen
  */
-class GitLab_API extends API {
+class GitLab_API extends API implements API_Interface {
 
 	/**
 	 * Holds loose class method name.
@@ -61,7 +61,7 @@ class GitLab_API extends API {
 	/**
 	 * Read the remote file and parse headers.
 	 *
-	 * @param $file
+	 * @param string $file Filename.
 	 *
 	 * @return bool
 	 */
@@ -132,7 +132,7 @@ class GitLab_API extends API {
 	/**
 	 * Read the remote CHANGES.md file.
 	 *
-	 * @param $changes
+	 * @param string $changes Changelog filename.
 	 *
 	 * @return bool
 	 */
@@ -357,12 +357,12 @@ class GitLab_API extends API {
 	/**
 	 * Add appropriate access token to endpoint.
 	 *
-	 * @param $git
-	 * @param $endpoint
+	 * @param object $git
+	 * @param string $endpoint
 	 *
 	 * @access private
 	 *
-	 * @return string
+	 * @return string $endpoint
 	 */
 	private function add_access_token_endpoint( $git, $endpoint ) {
 		// This will return if checking during shiny updates.
@@ -395,12 +395,12 @@ class GitLab_API extends API {
 	/**
 	 * Create GitLab API endpoints.
 	 *
-	 * @param $git      object
-	 * @param $endpoint string
+	 * @param object $git
+	 * @param string $endpoint
 	 *
-	 * @return string
+	 * @return string $endpoint
 	 */
-	protected function add_endpoints( $git, $endpoint ) {
+	public function add_endpoints( $git, $endpoint ) {
 
 		switch ( $git::$method ) {
 			case 'projects':
@@ -472,7 +472,7 @@ class GitLab_API extends API {
 	 *
 	 * @return object|array Array of tag numbers, object is error.
 	 */
-	protected function parse_tag_response( $response ) {
+	public function parse_tag_response( $response ) {
 		if ( isset( $response->message ) ) {
 			return $response;
 		}
@@ -494,7 +494,7 @@ class GitLab_API extends API {
 	 *
 	 * @return array $arr Array of meta variables.
 	 */
-	protected function parse_meta_response( $response ) {
+	public function parse_meta_response( $response ) {
 		$arr      = array();
 		$response = array( $response );
 
@@ -516,7 +516,7 @@ class GitLab_API extends API {
 	 *
 	 * @return array|object $arr Array of changes in base64, object if error.
 	 */
-	protected function parse_changelog_response( $response ) {
+	public function parse_changelog_response( $response ) {
 		if ( isset( $response->messages ) ) {
 			return $response;
 		}